+++ Examples/example1_demand_patterns.py
@@ -11,10 +11,10 @@ Created on Sat Oct  8 16:36:38 2022
 
 @author: Tyler
 """
-import visnetwork as vis
+import viswaternet as vis
 import matplotlib.pyplot as plt
 
-model = vis.VisnetModel(r'Networks/CTown.inp')
+model = vis.VisWNModel(r'Networks/CTown.inp')
 
 fig, ax = plt.subplots(figsize=(12,12))