+++ Examples/example2b_diameter_greatherthan.py
@@ -5,11 +5,11 @@ Created on Sat Oct 22 16:03:53 2022
 @author: Tyler
 """
 
-import visnetwork as vis
+import viswaternet as vis
 import matplotlib.pyplot as plt
 
 # Run EPANET2.0 Simulation and store results
-model = vis.VisnetModel(r"Networks/CTown.inp")  
+model = vis.VisWNModel(r"Networks/CTown.inp")  
 
 # Define figure to be drawn to
 fig, ax = plt.subplots(figsize=(12, 12))