+++ viswaternet/drawing/continuous.py
@@ -7,9 +7,9 @@ Created on Sun Oct  2 21:05:57 2022
 import numpy as np
 import matplotlib as mpl
 import networkx.drawing.nx_pylab as nxp
-from visnetwork.network import processing
-from visnetwork.utils import save_fig, normalize_parameter, unit_conversion
-from visnetwork.drawing import base
+from viswaternet.network import processing
+from viswaternet.utils import save_fig, normalize_parameter, unit_conversion
+from viswaternet.drawing import base
 
 
 default_cmap = mpl.cm.get_cmap("autumn_r")