+++ viswaternet/network/__init__.py
@@ -1,2 +1,2 @@
-from .initialize import VisnetModel
+from .initialize import VisWNModel
 from .processing import bin_parameter, get_parameter, get_demand_patterns