+++ HISTORY.rst
@@ -1,13 +1,20 @@
 =======
 History
 =======
+0.1.2 (2022-11-11)
+------------------
+
+* Rename package once more to avoid confusion with R package 'visnetwork'.
+
+0.1.1 (2022-11-11)
+------------------
+
+* Fix import errors due to package name change.
 
 0.1.0 (2022-11-11)
 ------------------
 
 * First release on PyPI.
 
-0.1.1 (2022-11-11)
-------------------
 
-* Fix import errors due to package name change.
+